requirements.txt
paho-mqtt
schedule
pyserial